Unit 6
Finished the unit 6 questions including drawing the secret animal using the turtle package and list of points.

Imported turtle package to draw the animal from the lists of points given.

Added the lists of points: first, second

Added function draw_lines which receives a list of points (x, y) and a color eg: 'red', 'blue' etc. And prints lines between the points using the turtle package.

In main function setup the turtle screen, speed and default color. Added the draw call for the first list of points with the color 'red', added the draw call for the second list of points with the color 'blue'.